[修改] 优化

+++ src/main/java/com/ruoyi/common/utils/security/ShiroUtils.java
@@ -57,7 +57,7 @@ public class ShiroUtils {
     }
 
     public static Long getUserId() {
-        return getSysUser().getUserId().longValue();
+        return getSysUser().getUserId();
     }
 
     public static String getLoginName() {
